Disallow unescaped newlines in strings in ocamltests
Ocamltest syntax allows embedding newlines like so:

```
flags = "\
  -I thing1 -I thing2 \
  -I thing3 -I thing4 \
";
```

The consequences of leaving off a backslash (and thus constructing a
command line with an embedded newline) appear to be subtle and have not been
fully investigated. It is an established fact, however, that they can include
_wasting multiple solid workdays of developer time._

Since all the tests pass with this restriction, there is apparently no need to
allow actual newlines to occur in these attribute values. We should just scrap
the footgun. Right now.