Remove excess arguments from globals when strengthening
When we chase down the aliases, we may find that

```
A.t
```

is in fact

```
Foo[P:Bar][Q:Baz].Bar.t
```

and furthermore that `Foo` only takes `P` and not `Q`. This is fine and
expected, but if we then return a signature that includes
```
type t = Foo[P:Bar][Q:Baz].t
```
then if another module accesses the same type more precisely as
`Foo[P:Bar].t` we have trouble: `Foo[P:Bar].t` and `Foo[P:Bar][Q:Baz].t`
should be the same type (since `Foo[P:Bar]` and `Foo[P:Bar][Q:Baz]` are
identical) but the two `t`s have different manifests so it's assumed that
they're different.

The solution is simple: rather than blindly using the path we looked up as the
path at which to strengthen the signature, normalize the path first. Since we've
necessarily loaded the .cmi at this point we know we can get the exact global
(that is, no excess arguments).