flambda2-types: Recover aliases after meet
The function `TG.recover_some_aliases` is called by the typing
environment in order to detect aliases prior to adding equations.

With the advanced meet (and also sometimes with the basic meet), this
alias recovery is done before computing the meet with the existing type
of the variable, which means that we can lose aliases if the meet
discovers a new alias.

This patch changes the logic to perfom alias recovery after the meet has
been performed, and also gives more descriptive names to the different
steps of `add_equation` to make the logic more clear.

If we have:

```
x: (Variant (blocks {tag_0}) (tagged_imms ((= #0))))
```

and we add:

```
x: (Variant (blocks ⊥) (tagged_imms ⊤))
```

without this patch we get:

```
x: (Variant (blocks ⊥) (tagged_imms (= #0)))
```

with this patch we get:

```
x: (= 0)
```