flambda2-types: Make meet return a new typing env
Make the `meet` function return a new typing env rather than an
extension. This is preparatory work for the new join algorithm,
which might create new variables (`meet` can call `join`), but env
extensions cannot hold variables.

An alternative would be to return env extensions with new variables, but
we are internally building a new typing env anyways -- it doesn't make
much sense to cut it then immediately add it back to the same env, so we
might as well just return the typing env directly.