flambda2-types: It is OK to add an alias between constants
We have an invariant that asserts we cannot add aliases between
constants, but we actually support it (by returning `Bottom`).

If we have `y` with an env extension on tag 0 that states `x` has type
`= 0`, and we later perform a switch on `x = 1` then `Get_tag y = 0`,
this would add the env extension to the environment and we want to
deduce this is an impossible case from the equation `0 = 1`.