Add a flag to control inference of with-bounds on type decls (#3547)
Add a new flag, -infer-with-bounds, which enables inference of
with-bounds (called "baggage" as of this commit, but to be renamed to
with-bounds later) on type declarations. This defaults to false, and has to be
turned on per-compilation-unit, because we are concerned about the performance
impact of rolling out with-kind inference and want to do so gradually.

If the flag is off, we restore the original inferred kinds prior to the
introduction of with-bounds.