Check annotations in a separate pass from updating jkinds
Mutually recursive types need to see each others' types during normalization, in
order to get the correct kind, and so need an env with all the types in a single
type declaration present, fully specified and marked as `best`. This was handled
correctly in normalize_decls_jkind, but not correctly (with /all/ decls in an
env together) when checking user-written /annotations/ against the jkind (which
is done in update_decl_jkind). This commit updates that to happen in a separate
pass after update_decl_jkind is called on all types, and adds a test that was
originally failing without the change.