Per-axis jkind errors
This PR adds a section to subjkind errors when either kind's with-bounds
requires a modality to be printed. The section lists each axis along which there
was disagreement, and what the jkinds are along those axes. The section is only
printed when the layouts extension is in alpha in order to match the behavior of
printing with-bounds. (If no with-bounds are ever printed, then no modalities
ever are.)