Fix checking of decl jkinds against original jkinds
this has some finnicky bits around checking layouts, since that has the side
effect of filling in sort variables (which is necessary, since otherwise we end
up defaulting to value later and getting the wrong jkind). But also we now check
the jkind after normalizing, which is nice since we probably have to do a lot
less work.