jkind subsumption (#3525)
Improve the implementation of Jkind.sub_jkind_l to give a proper subsumption
check. The implementation is based on the rules in
jane/doc/proposals/kind-inference.md. At the moment, this check is quadratic
because we don't have a proper comparison function on types. But based on my
anecdotal experience building and running tests, this doesn't seem to hurt
performance too bad.

Co-authored-by: Aspen Smith <aspsmith@janestreet.com>
Co-authored-by: Richard Eisenberg <reisenberg@janestreet.com>