flambda2: Implement a datalog engine for the reaper (#3476)
The reaper is essentially a (backwards) points-to analysis, a problem on which Datalog shines. Implement a small Datalog engine in order to be able to facilitate tweaking and extending the analysis (which we are doing as part of the unbox-free-vars-of-closures work).

This patch can be decomposed in roughly two parts: the implementation of the Datalog engine itself, and its use to supplant the existing reaper analysis. The old analysis is kept around in order to cross-check the results from the new Datalog analysis and validate its performance (currently the Datalog-based solver is around 2-2.5x faster than the previous solver).

The datalog engine itself lives in the new `Flambda_datalog` module, and is implemented on top of the Patricia tree module and the Leapfrog Triejoin [1] algorithm. This has been chosen instead of implementations based on hash-tables because it allows making the datalog database fully persistent, which can be useful for some potential future applications.

*Note*: The Datalog engine is intended to be usable for other applications; see examples in the `datalog` API test. It has however only been tested extensively in the context of the reaper -- please get in touch before using it for other applications.

[1] : Leapfrog Triejoin: a worst-case optimal join algorithm, Todd L. Veldhuizen (https://arxiv.org/abs/1210.0481) 

Co-authored-by: Nathanaëlle Courant <nathanaelle.courant@ocamlpro.com>