Skip irrelevant axes when we can during normalize
Skip normalizing along axes where we don't need to in as many places as
possible, by replacing normalize with map_normalize with a function that does an
axis set intersection with the axes we know we care about. This improves
performance of eg compiling typecore by a few percent, and is cheap to do.