build: Only set OCAMLPARAM if BUILD_OCAMLPARAM is defined (#3594)
Setting the value of OCAMLPARAM in the workspace file is useful for CI,
but annoying for local development as OCAMLPARAM is a good way to
selectively enable features when testing.

This patch sets the value of OCAMLPARAM while building the stdlib or
compiler to the value of BUILD_OCAMLPARAM, unless BUILD_OCAMLPARAM is
not defined.