Avoid traversing C parts of frame pointer chain when reallocating stack (#3376)
* Avoid traversing C parts of frame pointer chain when reallocating stack

When the OCaml stack grows we need to rewrite frame pointers (if enabled)
to point to the new stack.

However, when using a C library that was not compiled with frame pointers
enabled, we cannot assume that there is an unbroken chain of frame pointers
through both the OCaml and C parts of the stack. Doing so leads to segfaults.

Instead, we note that the only frame pointers that can point to OCaml stacks
(the ones that need updating) are those already on OCaml stacks, plus the first
ones pushed after any OCaml->C calls. These can be found by traversing the
struct c_stack_link chain, without needing to traverse any intervening C frames.

This imposes a new constraint on the runtime assembly stubs: after switching to
C they must not push anything to the stack before calling a C function. This
was already true for all but caml_c_call_stack_args. Enforcing this invariant
for caml_c_call_stack_args is straightforward enough, and simplifies the DWARF
backtrace logic.

For arm64, a side-effect of this change is that DWARF backtraces now work on
stacks containing calls to caml_c_call_stack_args, which were broken before.
(Tested with macos lldb)