Add parts of aliased instance names to imports
If the user writes

    module Foo = Bar(Param)(Arg) [@jane.non_erasable.instances]

under `-no-alias-deps`, then we add `Bar` to the list of interface imports (with
no CRC), but not `Param` or `Arg`. This doesn't really reflect what's going on:
resolving that alias will require loading all three of `bar.cmi`, `param.cmi`,
and `arg.cmi`. As a result, `ocamlobjinfo` is giving incomplete information
about dependencies.