Add more information to `ocamlobjinfo` output (#3622)
* Add objinfo tests to `tests/templates/dunelike`

* Add parts of aliased instance names to imports

If the user writes

    module Foo = Bar(Param)(Arg) [@jane.non_erasable.instances]

under `-no-alias-deps`, then we add `Bar` to the list of interface imports (with
no CRC), but not `Param` or `Arg`. This doesn't really reflect what's going on:
resolving that alias will require loading all three of `bar.cmi`, `param.cmi`,
and `arg.cmi`. As a result, `ocamlobjinfo` is giving incomplete information
about dependencies.

* objinfo: Print arguments of imported instances

We've been printing the `Import_info.name`s of implementation imports, but this
leaves off the arguments of any instances that are imported (often leading to
multiple rows with the same name but different CRCs).

* Add a test to make sure `test_{byte,native}.ml` remain in sync

* Fix tests by blanking out `OCAMLPARAM`

* Make test fail explicitly in error case