Define lattice ops in terms of conversion to an integer
From some digging around in godbolt, this generates *way* nicer code than the
explicit match - plus it's (imo) clearer and easier to get right.

This does not, in the current state of the world, show up on a quick benchmark,
but the generated assembly is so obviously better - and the code not obviously
worse (and arguably nicer) that I somewhat think we should make this change
anyway.