Fix bug in lazy pattern matching
In some situations (i.e. with specialization), flambda2 might propagate
too much information. In the situation of a branch where a lazy (whose
definition by a make_block is seen by flambda2) is shortcut, and then
directly used as an integer (because the lazy wraps an enum/variant),
flambda2 might consider the branch as dead code, since the result of a
make_block (the lazy value) is used as an integer. This is obviously not
correct since the lazy value can be shortcut. The fix is to add an
(obvisouly ?) missing opaque when creating a lazy block.

Additionally, the let-rec static size computation needs some updating.