Keep track of relevant axes when descending in normalize (#3670)
When we descend through types in normalize, we need to keep track of a
successive intersection of which axes we actually care about (eg if some are
skipped because of modalities). This adds a new `relevant_axes` argument to loop
which tracks that, and only does the join along the intersection of those
relevant axes and the relevant axes in the type_info.

This fixes a bug where we didn't actually mode cross when a type which doesn't
normally mode cross along that axis was nested inside a modality which should
have allowed crossing.