Fix bug in lazy pattern matching (#3647)
* Add new primitives to create blocks for lazy values

This adds a new primitive in Lambda and Flambda2 to create a block
for a (not short-circuited) lazy value. These primitives allow to
more systematically handle lazy blocks correctly, for instance with
respect to the value-let-rec size computation, and in flambda to better
ensure that we do not track block appromixations for lazy values
(without relying on the frontend to insert opaque identities).

* cleanup

* Update code size of Make_lazy to match Make_block

* Remove allocation mode from lazy primitives

* Formatting

---------

Co-authored-by: Vincent Laviron <vincent.laviron@gmail.com>