Introduce `Global_module.Parameter_name.t`
Takes the place of any `Global_module.t` or `Global_module.Name.t` that is known
to be the name of a parameter. Can't have parameters. (Notably, even the current
half-worked-out plans for parameterized parameters would actually keep this type
in many places, so this isn't as much in the way as one might think.)