Remove all objects from ocamltest (#3821)
A simple refactor. Replaces all objects with first-class modules. Inheritance is
mimicked using `include`. There is exactly once instance where that doesn't
work, since a method gets overridden and one one of the remaining methods calls
it, but all that's necessary is to “override” both “methods” (and they're not
even in different files so it's easy).

The formatting is pretty wonky as I was hoping to keep a small diff in case
changes are made upstream. (Obviously this can only help so much, especially
given the repeated uses of `method`.)