Generate and propagate never returns from emit_tuple
In some cases, particularly in dead code (which happens to be a bit more
easily to produce with the upcoming match-in-match), `to_cmm` can
produce code that makes Selection fails, because emit_tuple refuses to
emit code that never returns. This PR simply allows emit_tuple to return
`Never_returns`, and the required change to propagate that value upwards
until it can be handled.