Remove unnecessary submode check in [type_omitted_parameters]
The check looks important, but I don't think it is:
* It's not actually always run -- the first several arguments to a
function, before any omitted arguments, aren't checked anyway. So if
this *is* important, then I think we have a bigger problem.
* If the partially-applied function call type-checks, then any
non-bottom (comonadic) mode on an argument must be reflected in the
type of the function itself, and so checks around [mode_fun] will be
sufficient.