Fix name abstraction when simplifying switch
When simplifying a switch expression, we introduce a new variable to
represent the tagged version of the scrutinee, which is a naked
immediate.

This variable is (presumably) added only so that it can get picked up by
CSE when performing simplifications in the cases where a tagged version
of the scrutinee is not otherwise available (e.g. the scrutinee comes
from an `Is_int` or `Get_tag` primitive), and is marked as
`used_in_current_handler` for this purpose.

However the actual processing of this variable is delegated to
`simplify_let`, which immediately opens its name abstraction -- causing
a desynchronisation between the variable that we recorded as being used
and the one that actually ends up in the CSE equations / the typing env.

This patch introduces a new `simplify_let_with_bound_pattern` that
forwards the `Bound_pattern.t` after opening the name abstraction and
uses it to record the proper variable as used, causing it to appear in
the `required_names` and allowing simplifications to proceed.

== Concrete changes

Before this patch, functions such as `Option.is_none` and
`Option.is_some` get compiled to a switch. After this patch, they get
compiled to `Tag_imm (Is_int .)` and `Boolean_not (Tag_imm (Is_int .))`,
respectively.

Functions that manually compute the tag of a block by returning the
appropriate integer also now get compiled to `Tag_imm (Get_tag .)`
earlier (before this patch it was done by `To_cmm`, after this patch it
happens inside flambda2).