Use `Unit_info.t` more like upstream (#3926)
* Add `kind` to `Unit_info.t`

* Use `Compilation_unit.t` in place of `modname` in `Unit_info`

This requires knowing the pack prefix wherever a `Unit_info.t` is created, but
generally we're creating a `Compilation_unit.t` and a `Unit_info.t` at the same
time so the caller already knows the pack prefix.

* Keep the current unit as a `Unit_info.t`, not a `Compilation_unit.t`

This makes downstream and upstream much better aligned, since upstream uses
`Unit_info.t` in this way. (Some of the changes here are just copied from
upstream, in fact.) There's a bit of awkwardness in that we often need to use
dummy unit info because (unlike upstream) our typechecker can't run without a
current `Compilation_unit.t` set.