flambda2: Continuation shortcuts (#3545)
Replace the continuation alias mechanism with a more general
continuation shortcut mechanism.

A continuation shortcut is a let_cont whose body is an arbitrary
apply_cont. A continuation alias is a special case of continuation
shortcuts where the arguments of the apply_cont are exactly the
parameters of the let_cont.

Shortcuts are detected and resolved in the upwards pass after rewriting
the apply_cont, so that we should never output code containing
continuation shortcuts (except when used as return or exception
continuation).

This should help generate simpler control flow for later Simplify or
To_cmm passes, and make the output of `-dflambda` easier to follow.

Fixes #3210