Have errors suggest "shared or uncontended" rather than "shared" (#3791)
The error message
```
Error: This value is contended but expected to be shared.
```
is unhelpful for users that know about `uncontended` but not `shared` (which
will likely be many users if not most). In particular, it won't be obvious to
them that it suffices to make the value `uncontended`. This changes the
message to
```
Error: This value is contended but expected to be shared or uncontended.
```
which hopefully is useful even if the word `shared` is mysterious.