Make the AST version mismatch error message less cryptic
Today I spent nearly 2 hours trying to understand why I was failing to build our internal
codebase with an older build of the compiler. Ultimately the issue was that the magic
number hard-coded in `ppxlib/astlib/ast_99.ml` didn't match the magic number of the
compiler (even though in this case the parse-trees were still compatible). To save
others the pain of having to figure this out in the future, I've made the error message
produced in this scenario more descriptive.