Make all atomic primitives take a field index
To pave the road to atomic record fields, make all atomic primitives from
lambda onwards operate on an arbitrary field index of a block, rather than
always operating on the first field. The previous primitives (%atomic_load, etc)
are left over for upstream compatibility, but are translated to field-based
operations hardcoded to field 0 as soon as they hit lambda.

Some of this was taken from the upstream PR[0] for atomic fields, but diverges
significantly from that in that upstream doesn't have flambda2, and also in that
upstream only has C stubs for atomic primitives, not codegen.

[0] https://github.com/ocaml/ocaml/pull/13404